+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [3.4.0] - 2018-05-23
+### Added
+ - Added SetEquals method to IcdHashSet
+
 ## [3.3.0] - 2018-05-18
 ### Added
  - Added IcdOrderedDictionary collection